Implemented SaveDataRedirector - namespace redirection
By directly mixin to ResourceLocation's constructor. This should not significantly cause lags. Upon testing, during loading mods, ResourceLocations are created 100,000 times each 0.5s at most frequency (when loading game or save data), and reading a map for 100,000 times is not extremely costly.

The redirector will only available in Java code but not data, as it's designed for only registry key changes or modid changes, but not any game mechanic changes.